requirements.txt
torch==2.2.0
torchvision==0.17.0
timm==0.9.16
opencv-python-headless==4.9.0.80
numpy==1.26.4
pandas==2.2.1
matplotlib==3.8.3
seaborn==0.13.2
scikit-learn==1.4.1.post1
Pillow==10.2.0
tqdm==4.66.2
albumentations==1.4.3
streamlit==1.32.0
google-genai>=0.7.0
